♻️ [Refactor] 프론트 script 모듈로 분리
