+++ public/scripts/eventListeners.js
@@ -1,3 +1,30 @@
+import { state } from "./socket.js";
+import {
+  loginApi,
+  loginNodeApi,
+  getFriendListApi,
+  getChatroomListApi,
+  enterChatroomApi,
+  getMessageApi,
+  startChatByMemberIdApi,
+  startChatByBoardIdApi,
+  starFriendApi,
+} from "./api.js";
+
+import { renderChatroomDiv } from "./module.js";
+
+const loginButton = document.getElementById("loginButton");
+const fetchFriendsButton = document.getElementById("fetchFriendsButton");
+const logoutButton = document.getElementById("logoutButton");
+const fetchChatroomsButton = document.getElementById("fetchChatroomsButton");
+const notificationButton = document.getElementById("notificationButton");
+const notiCloseButton = document.getElementById("notiCloseButton");
+const boardTestButton = document.getElementById("boardTestButton");
+const closeBoardIdPopupButton = document.getElementById("closeBoardIdPopupButton");
+const boardIdForm = document.getElementById("boardIdForm");
+const matchingTestButton = document.getElementById("matchingTestButton");
+const closeMatchingMemberIdPopupButton = document.getElementById("closeMatchingMemberIdPopupButton");
+const mathingMemberIdForm = document.getElementById("mathingMemberIdForm");
 const userEmailInput = document.getElementById("userEmail");
 const userPwInput = document.getElementById("userPw");
 
@@ -26,10 +53,9 @@ loginButton.addEventListener("click", () => {
       localStorage.setItem("refreshToken", refreshToken);
 
       alert("Login successful! Token received.");
-      console.log("Using existing socket. Socket ID:", socket.id);
 
       // (#1-5) 3000서버로 login api 요청 (소켓 초기화를 위함)
-      loginNodeApi();
+      loginNodeApi(state.socketId);
     }
   });
 });
@@ -49,15 +75,12 @@ fetchFriendsButton.addEventListener("click", () => {
       const friendsElement = document.getElementById("friendsList");
       friendsElement.innerHTML = ""; // 이전 친구목록 element 비우기
 
-      hasNextFriend = result.has_next;
-      nextFriendCursor = result.next_cursor;
-
       // (#3-3) 친구 목록 화면 렌더링
-      result.friendInfoDTOList.forEach((friend) => {
+      result.friendInfoList.forEach((friend) => {
         const li = document.createElement("li");
 
         // onlineFriendMemberIdList에 존재하는 회원인지 (해당 회원이 현재 온라인인지)에 따라 상태 text 배정
-        const isOnline = onlineFriendMemberIdList.includes(friend.memberId);
+        const isOnline = state.onlineFriendMemberIdList.includes(friend.memberId);
         const statusText = isOnline ? "online" : "offline";
 
         // 접속 상태 element 생성 및 class 부여
@@ -71,7 +94,7 @@ fetchFriendsButton.addEventListener("click", () => {
         const userInfoDiv = document.createElement("div");
         userInfoDiv.className = "user-info"; // CSS 클래스 추가
         userInfoDiv.innerHTML = `
-                <img src="${friend.memberProfileImg}" alt="${friend.name}'s profile picture" width="30" height="30">
+                <img src="${friend.profileImg}" alt="${friend.name}'s profile picture" width="30" height="30">
                 <span>${friend.name}</span>
             `;
         li.appendChild(userInfoDiv);
@@ -96,53 +119,47 @@ fetchFriendsButton.addEventListener("click", () => {
           // (#13-1) 채팅방 시작 API 요청
           startChatByMemberIdApi(friend.memberId).then((result) => {
             // (#13-2) 채팅방 시작 API 정상 응답 받음
-            if (result.chatMessageList) {
+            if (result.chatMessageListResponse) {
               // 채팅 메시지 내역이 있을 때에만
               // (#13-3) messagesFromThisChatroom array 초기화
-              messagesFromThisChatroom = result.chatMessageList.chatMessageDtoList;
+              state.messagesFromThisChatroom = result.chatMessageListResponse.chatMessageList;
 
               console.log("============== fetch chat messages result ===============");
-              console.log(result.chatMessageList.chatMessageDtoList);
+              console.log(result.chatMessageListResponse.chatMessageList);
 
               // (#13-4) hasNextChat 업데이트
-              hasNextChat = result.chatMessageList.has_next;
+              state.hasNextChat = result.chatMessageListResponse.hasNext;
             } else {
               // 채팅 메시지 내역이 아예 없을 경우
               // messagesFromThisChatroom array 초기화
-              messagesFromThisChatroom = [];
+              state.messagesFromThisChatroom = [];
 
-              // hasNext null로 초기화
-              hasNextChat = null;
+              // (#13-4) hasNextChat 업데이트
+              state.hasNextChat = null;
             }
 
             // (#13-5) 현재 보고 있는 채팅방 uuid, 채팅 중인 memberId 업데이트
-            currentViewingChatroomUuid = result.uuid;
-            currentChattingMemberId = result.memberId;
+            state.currentViewingChatroomUuid = result.uuid;
+            state.currentChattingMemberId = result.memberId;
 
             // (#13-6) systemFlag 초기화, 특정 회원과의 채팅방 시작 시 systemFlag는 항상 null임
-            currentSystemFlag = null;
+            state.currentSystemFlag = null;
 
-            renderChatroomDiv(result);
+            renderChatroomDiv(state, result);
           });
         });
 
         // 별 아이콘 클릭 시, 해당 친구 즐겨찾기 설정/해제 요청 전송
         star.addEventListener("click", () => {
+          starFriendApi(friend.memberId);
+
           // liked 클래스가 있는지 확인
           if (star.classList.contains("liked")) {
-            // 즐겨찾기 해제 api 요청
-            unstarFriendApi(friend.memberId).then((result) => {
-              // liked 상태 변경
-              star.classList.remove("liked");
-              console.log("즐겨찾기 해제 성공, memberId: ", friend.memberId);
-            });
+            // liked 상태 변경
+            star.classList.remove("liked");
           } else {
-            // 즐겨찾기 설정 api 요청
-            starFriendApi(friend.memberId).then((result) => {
-              // liked 상태 변경
-              star.classList.add("liked");
-              console.log("즐겨찾기 설정 성공, memberId: ", friend.memberId);
-            });
+            // liked 상태 변경
+            star.classList.add("liked");
           }
         });
 
@@ -151,131 +168,6 @@ fetchFriendsButton.addEventListener("click", () => {
     }
   });
 });
-
-// 친구 목록 내부에서 스크롤이 가장 아래에 닿았을 때
-document.addEventListener("DOMContentLoaded", () => {
-  const friendsListElement = document.getElementById("friendsList");
-
-  friendsListElement.addEventListener("scroll", () => {
-    // 스크롤이 가장 아래에 닿았는지 확인
-    const isScrollAtBottom = friendsListElement.scrollTop + friendsListElement.clientHeight >= friendsListElement.scrollHeight;
-
-    if (isScrollAtBottom) {
-      // 더 조회해올 다음 친구 목록이 있다면
-      if (hasNextFriend) {
-        fetchNextFriends(nextFriendCursor);
-      } else {
-        console.log("=== End of Friend List, No fetch ===");
-      }
-    }
-  });
-});
-
-// nextFriendCursor 기반으로 다음 친구 목록 조회 api 요청 및 화면 렌더링
-function fetchNextFriends(cursor) {
-  const jwtToken = localStorage.getItem("jwtToken");
-  if (!jwtToken) {
-    console.error("JWT token is missing.");
-    return;
-  }
-
-  getFriendListApi(cursor).then((result) => {
-    if (result) {
-      // (#3-2) 친구 목록 조회 성공 응답 받음
-      const friendsElement = document.getElementById("friendsList");
-
-      hasNextFriend = result.has_next;
-      nextFriendCursor = result.next_cursor;
-
-      // (#3-3) 친구 목록 화면 렌더링
-      result.friendInfoDTOList.forEach((friend) => {
-        const li = document.createElement("li");
-
-        // onlineFriendMemberIdList에 존재하는 회원인지 (해당 회원이 현재 온라인인지)에 따라 상태 text 배정
-        const isOnline = onlineFriendMemberIdList.includes(friend.memberId);
-        const statusText = isOnline ? "online" : "offline";
-
-        // 접속 상태 element 생성 및 class 부여
-        const statusElement = document.createElement("span");
-        statusElement.textContent = `(${statusText})`;
-        statusElement.classList.add(isOnline ? "online" : "offline");
-
-        li.setAttribute("data-member-id", friend.memberId);
-
-        // 사용자 정보를 담는 div 생성
-        const userInfoDiv = document.createElement("div");
-        userInfoDiv.className = "user-info"; // CSS 클래스 추가
-        userInfoDiv.innerHTML = `
-                <img src="${friend.memberProfileImg}" alt="${friend.name}'s profile picture" width="30" height="30">
-                <span>${friend.name}</span>
-            `;
-        li.appendChild(userInfoDiv);
-        li.appendChild(statusElement);
-
-        // 즐겨찾기 여부 표시
-        // 별 아이콘 추가
-        const star = document.createElement("span");
-        star.className = "star";
-        star.innerHTML = "☆";
-
-        // liked가 true인 경우 보라색으로 설정
-        if (friend.liked) {
-          star.classList.add("liked");
-        }
-
-        // li 요소에 별 아이콘 추가
-        li.appendChild(star);
-
-        // 해당 친구 부분 클릭 시, 친구와의 채팅방 시작 api 요청
-        userInfoDiv.addEventListener("click", () => {
-          // (#13-1) 채팅방 시작 API 요청
-          startChatByMemberIdApi(friend.memberId).then((result) => {
-            // (#13-2) 채팅방 시작 API 정상 응답 받음
-            // (#13-3) messagesFromThisChatroom array 초기화
-            messagesFromThisChatroom = result.chatMessageList.chatMessageDtoList;
-
-            console.log("============== fetch chat messages result ===============");
-            console.log(result.chatMessageList.chatMessageDtoList);
-
-            // (#13-4) hasNextChat 업데이트
-            hasNextChat = result.chatMessageList.has_next;
-
-            // (#13-5) 현재 보고 있는 채팅방 uuid, 채팅 중인 memberId 업데이트
-            currentViewingChatroomUuid = result.uuid;
-            currentChattingMemberId = result.memberId;
-
-            // (#13-6) systemFlag 초기화, 특정 회원과의 채팅방 시작 시 systemFlag는 항상 null임
-            currentSystemFlag = null;
-
-            renderChatroomDiv(result);
-          });
-        });
-
-        // 별 아이콘 클릭 시, 해당 친구 즐겨찾기 설정/해제 요청 전송
-        star.addEventListener("click", () => {
-          // liked 클래스가 있는지 확인
-          if (star.classList.contains("liked")) {
-            // 즐겨찾기 해제 api 요청
-            unstarFriendApi(friend.memberId).then((result) => {
-              // liked 상태 변경
-              star.classList.remove("liked");
-              console.log("즐겨찾기 해제 성공, memberId: ", friend.memberId);
-            });
-          } else {
-            // 즐겨찾기 설정 api 요청
-            starFriendApi(friend.memberId).then((result) => {
-              // liked 상태 변경
-              star.classList.add("liked");
-              console.log("즐겨찾기 설정 성공, memberId: ", friend.memberId);
-            });
-          }
-        });
-
-        friendsElement.appendChild(li);
-      });
-    }
-  });
-}
 
 // 로그아웃 버튼 클릭 시
 logoutButton.addEventListener("click", () => {
@@ -293,7 +185,7 @@ logoutButton.addEventListener("click", () => {
     headers: {
       "Content-Type": "application/json",
       Authorization: `Bearer ${jwtToken}`,
-      "Socket-Id": socket.id,
+      "Socket-Id": state.socketId,
     },
   })
     .then((response) => response.json())
@@ -320,19 +212,16 @@ fetchChatroomsButton.addEventListener("click", () => {
   }
 
   // (#8-1) 채팅방 목록 조회 api 요청
-  getChatroomListApi(null).then((result) => {
+  getChatroomListApi().then((result) => {
     if (result) {
       // (#8-2) 채팅방 목록 조회 성공 응답 받음
       // 채팅방 목록 element 초기화
       const chatroomListElement = document.getElementById("chatroomList");
       chatroomListElement.innerHTML = "";
 
-      hasNextChatroom = result.has_next;
-      nextChatroomCursor = result.next_cursor;
-
       // (#8-3) 채팅방 목록 렌더링
       // api result data를 돌면서 html 요소 생성
-      result.chatroomViewDTOList.forEach((chatroom) => {
+      result.chatroomResponseList.forEach((chatroom) => {
         const li = document.createElement("li");
         li.classList.add("chatroom-item");
         li.setAttribute("data-chatroom-uuid", chatroom.uuid); // data-chatroom-uuid 값 세팅
@@ -365,78 +254,6 @@ fetchChatroomsButton.addEventListener("click", () => {
     }
   });
 });
-
-// 채팅방 목록 내부에서 스크롤이 가장 아래에 닿았을 때
-document.addEventListener("DOMContentLoaded", () => {
-  const chatroomListElement = document.getElementById("chatroomList");
-
-  chatroomListElement.addEventListener("scroll", () => {
-    // 스크롤이 가장 아래에 닿았는지 확인
-    const isScrollAtBottom = chatroomListElement.scrollTop + chatroomListElement.clientHeight >= chatroomListElement.scrollHeight;
-
-    if (isScrollAtBottom) {
-      // 더 조회해올 다음 친구 목록이 있다면
-      if (hasNextChatroom) {
-        fetchNextChatrooms(nextChatroomCursor);
-      } else {
-        console.log("=== End of Chatroom List, No fetch ===");
-      }
-    }
-  });
-});
-
-// nextChatroomCursor 기반으로 다음 채팅방 목록 조회 api 요청 및 화면 렌더링
-function fetchNextChatrooms(cursor) {
-  const jwtToken = localStorage.getItem("jwtToken");
-  if (!jwtToken) {
-    console.error("JWT token is missing.");
-    return;
-  }
-
-  getChatroomListApi(cursor).then((result) => {
-    if (result) {
-      // (#8-2) 채팅방 목록 조회 성공 응답 받음
-      // 채팅방 목록 element 초기화
-      const chatroomListElement = document.getElementById("chatroomList");
-
-      hasNextChatroom = result.has_next;
-      nextChatroomCursor = result.next_cursor;
-
-      // (#8-3) 채팅방 목록 렌더링
-      // api result data를 돌면서 html 요소 생성
-      result.chatroomViewDTOList.forEach((chatroom) => {
-        const li = document.createElement("li");
-        li.classList.add("chatroom-item");
-        li.setAttribute("data-chatroom-uuid", chatroom.uuid); // data-chatroom-uuid 값 세팅
-
-        li.innerHTML = `
-                                                            <div>
-                                                                <img src="${chatroom.targetMemberImg}" alt="Profile Image" width="30" height="30">
-                                                            </div>
-                                                            <div class="chatroom-info">
-                                                                <span>${chatroom.targetMemberName}</span>
-                                                                <p last-msg-text>${chatroom.lastMsg ? chatroom.lastMsg : " "}</p>
-                                                            </div>
-                                                            <div>
-                                                                <p last-msg-time>${new Date(chatroom.lastMsgAt).toLocaleString()}</p>
-                                                                <p data-new-count>${chatroom.notReadMsgCnt}</p>
-                                                                <button class="enter-chatroom-btn" data-chatroom-uuid="${chatroom.uuid}">채팅방 입장</button>
-                                                            </div>
-                                                        `;
-        chatroomListElement.appendChild(li);
-      });
-
-      // 채팅방 입장 버튼에 eventListener 추가
-      const enterChatroomButtons = document.querySelectorAll(".enter-chatroom-btn");
-      enterChatroomButtons.forEach((button) => {
-        button.addEventListener("click", (event) => {
-          const chatroomUuid = event.target.getAttribute("data-chatroom-uuid");
-          enterChatroom(chatroomUuid);
-        });
-      });
-    }
-  });
-}
 
 // 채팅방 입장 시
 function enterChatroom(chatroomUuid) {
@@ -453,26 +270,25 @@ function enterChatroom(chatroomUuid) {
       // (#9-2) 채팅방 입장 API 정상 응답 받음
 
       // (#9-3) messagesFromThisChatroom array 초기화
-      messagesFromThisChatroom = result.chatMessageList.chatMessageDtoList;
+      state.messagesFromThisChatroom = result.chatMessageListResponse.chatMessageList;
 
       console.log("============== fetch chat messages result ===============");
-      console.log(result.chatMessageList.chatMessageDtoList);
+      console.log(result.chatMessageListResponse.chatMessageList);
 
-      // (#9-4) hasNextChat 업데이트
-      hasNextChat = result.chatMessageList.has_next;
+      state.hasNextChat = result.chatMessageListResponse.hasNext;
 
       // (#9-5) 현재 보고 있는 채팅방 uuid, 채팅 중인 memberId 업데이트
-      currentViewingChatroomUuid = chatroomUuid;
-      currentChattingMemberId = result.memberId;
+      state.currentViewingChatroomUuid = chatroomUuid;
+      state.currentChattingMemberId = result.memberId;
 
       // (#9-6) systemFlag 업데이트
       // systemFlag update, 기존에 보던 채팅방과 다른 방에 입장한 경우에만
-      if (currentViewingChatroomUuid != chatroomUuid) {
-        currentSystemFlag = null; // uuid를 통한 채팅방 입장 시 systemFlag 값 없음
+      if (state.currentViewingChatroomUuid != chatroomUuid) {
+        state.currentSystemFlag = null; // uuid를 통한 채팅방 입장 시 systemFlag 값 없음
       }
 
       // 채팅방 영역 렌더링 메소드 호출
-      renderChatroomDiv(result);
+      renderChatroomDiv(state, result);
     }
   });
 }
@@ -485,7 +301,7 @@ document.addEventListener("DOMContentLoaded", () => {
     // 스크롤이 끝까지 올라갔을 때
     if (messagesElement.scrollTop === 0) {
       // 더 조회해올 다음 chat 내역이 있다면
-      if (hasNextChat) {
+      if (state.hasNextChat) {
         fetchOlderMessages();
       } else {
         console.log("=== End of Chat Messages, No fecth ==");
@@ -504,28 +320,28 @@ function fetchOlderMessages() {
 
   let firstMessage;
   // (#12-1) 가장 앞에 있는 메시지의 timestamp를 가져오기
-  if (messagesFromThisChatroom.length > 0) {
-    firstMessage = messagesFromThisChatroom[0];
+  if (state.messagesFromThisChatroom.length > 0) {
+    firstMessage = state.messagesFromThisChatroom[0];
   } else {
     console.log("messagesFromThisChatroom array is Empty");
   }
 
   // (#12-2) timestamp cursor를 담아 다음 메시지 조회 api 요청
-  getMessageApi(currentViewingChatroomUuid, firstMessage.timestamp).then((result) => {
+  getMessageApi(state.currentViewingChatroomUuid, firstMessage.timestamp).then((result) => {
     if (result) {
       // (#12-3) 다음 메시지 조회 성공 응답 받음
-      const olderMessages = result.chatMessageDtoList;
+      const olderMessages = result.chatMessageList;
 
       // (#12-4) messagesFromThisChatroom array 맨 앞에 새로 조회해온 orderMessages 추가
       // olderMessages 배열을 messagesFromThisChatroom 배열의 맨 앞에 추가
-      const updatedMessages = olderMessages.concat(messagesFromThisChatroom);
-      messagesFromThisChatroom = updatedMessages;
+      const updatedMessages = olderMessages.concat(state.messagesFromThisChatroom);
+      state.messagesFromThisChatroom = updatedMessages;
 
       console.log("===================== fetxh OlderMessages, updated messagesFromThisChatroom ==================");
-      console.log(messagesFromThisChatroom);
+      console.log(state.messagesFromThisChatroom);
 
       //(#12-5) hasNextChat 업데이트
-      hasNextChat = result.has_next;
+      state.hasNextChat = result.hasNext;
 
       // (#12-6) 기존 메시지 목록 앞에 새로 불러온 메시지 추가 렌더링
       const messagesElement = document.getElementById("messages");
@@ -553,7 +369,7 @@ function fetchOlderMessages() {
                   `;
         } else {
           // 시스템 메시지가 아닌 경우
-          if (message.senderId === memberId) {
+          if (message.senderId === state.memberId) {
             li.classList.add("mine");
           }
           li.innerHTML = `
@@ -576,35 +392,14 @@ function fetchOlderMessages() {
   });
 }
 
-// 채팅 전송 폼 제출 시
-form.addEventListener("submit", (e) => {
-  e.preventDefault();
-  if (input.value) {
-    const msg = input.value;
-    console.log("CHAT-MESSAGE EVENT EMIT ====== currentSystemFlag: ", currentSystemFlag);
-    if (currentSystemFlag) {
-      // 보내야 할 systemFlag가 있는 경우
-      // (#10-2) "chat-message" event emit
-      socket.emit("chat-message", { uuid: currentViewingChatroomUuid, message: msg, system: currentSystemFlag });
-
-      // systemFlag 초기화
-      currentSystemFlag = null;
-    } else {
-      // (#10-1) "chat-message" event emit
-      socket.emit("chat-message", { uuid: currentViewingChatroomUuid, message: msg });
-    }
-    input.value = "";
-  }
-});
-
 // 알림 아이콘 클릭 시
-document.getElementById("notificationButton").addEventListener("click", () => {
+notificationButton.addEventListener("click", () => {
   const notificationList = document.getElementById("notificationList");
   notificationList.style.display = notificationList.style.display === "block" ? "none" : "block";
 });
 
 // 알림 닫기 버튼 클릭 시
-document.getElementById("notiCloseButton").addEventListener("click", function () {
+notiCloseButton.addEventListener("click", function () {
   document.getElementById("notificationList").style.display = "none";
 });
 
@@ -623,33 +418,19 @@ document.querySelectorAll(".tab").forEach((tab) => {
   });
 });
 
-// 알림 요소 추가 메소드
-function addNotification(type, message, time, isRead) {
-  const notificationContainer = document.getElementById(`${type}Notifications`);
-  const div = document.createElement("div");
-  div.className = `notification-item ${isRead ? "read" : "unread"}`;
-  div.innerHTML = `
-    <div class="notification-text">
-      <p>${message}</p>
-      <p class="notification-time">${time}</p>
-    </div>
-  `;
-  notificationContainer.appendChild(div);
-}
-
 // 게시판 말걸어보기 테스트 버튼 클릭 시
-document.getElementById("boardTestButton").addEventListener("click", () => {
+boardTestButton.addEventListener("click", () => {
   document.getElementById("boardIdPopup").style.display = "block";
 });
 
 // 게시판 말걸어보기 팝업 닫기 버튼 클릭 시
-document.getElementById("closeBoardIdPopupButton").addEventListener("click", () => {
+closeBoardIdPopupButton.addEventListener("click", () => {
   document.getElementById("boardIdPopup").style.display = "none";
 });
 
 // 게시판 말걸어보기 팝업 제출 버튼 클릭 시
 // 게시글을 통한 채팅방 시작
-document.getElementById("boardIdForm").addEventListener("submit", (event) => {
+boardIdForm.addEventListener("submit", (event) => {
   event.preventDefault(); // 기본 제출 동작 방지
   document.getElementById("boardIdPopup").style.display = "none";
 
@@ -658,7 +439,7 @@ document.getElementById("boardIdForm").addEventListener("submit", (event) => {
   startChatByBoardIdApi(boardId).then((result) => {
     // (#16-2) 채팅방 시작 API 정상 응답 받음
     // (#16-3) messagesFromThisChatroom array 초기화
-    messagesFromThisChatroom = result.chatMessageList.chatMessageDtoList;
+    state.messagesFromThisChatroom = result.chatMessageListResponse.chatMessageList;
 
     // (#16-4) messagesFromThisChatroom 맨 뒤에 시스템 메시지 임의로 추가 (프론트 화면에서 보여지기 위함)
     if (result.system.flag === 1) {
@@ -671,7 +452,7 @@ document.getElementById("boardIdForm").addEventListener("submit", (event) => {
         timestamp: null,
         boardId: result.system.boardId,
       };
-      messagesFromThisChatroom.push(systemMessage);
+      state.messagesFromThisChatroom.push(systemMessage);
     } else {
       const systemMessage = {
         senderId: 0,
@@ -682,370 +463,43 @@ document.getElementById("boardIdForm").addEventListener("submit", (event) => {
         timestamp: null,
         boardId: result.system.boardId,
       };
-      messagesFromThisChatroom.push(systemMessage);
+      state.messagesFromThisChatroom.push(systemMessage);
     }
 
     console.log("============== fetch chat messages result ===============");
-    console.log(result.chatMessageList.chatMessageDtoList);
+    console.log(result.chatMessageListResponse.chatMessageList);
 
     // (#16-5) hasNextChat 업데이트
-    hasNextChat = result.chatMessageList.has_next;
+    state.hasNextChat = result.chatMessageListResponse.hasNext;
 
     // (#16-6) 현재 보고 있는 채팅방 uuid, 채팅 중인 memberId 업데이트
-    currentViewingChatroomUuid = result.uuid;
-    currentChattingMemberId = result.memberId;
+    state.currentViewingChatroomUuid = result.uuid;
+    state.currentChattingMemberId = result.memberId;
 
     // (#16-7) systemFlag 초기화
-    currentSystemFlag = result.system;
-    console.log("currentSystemFlag: ", currentSystemFlag);
+    state.currentSystemFlag = result.system;
+    console.log("currentSystemFlag: ", state.currentSystemFlag);
 
-    renderChatroomDiv(result);
+    renderChatroomDiv(state, result);
   });
 });
 
 // 매칭 채팅방 입장 테스트 버튼 클릭 시
-document.getElementById("matchingTestButton").addEventListener("click", () => {
+matchingTestButton.addEventListener("click", () => {
   document.getElementById("matchingMemberIdPopup").style.display = "block";
 });
 
 // 매칭 채팅방 입장 팝업 닫기 버튼 클릭 시
-document.getElementById("closeMatchingMemberIdPopupButton").addEventListener("click", () => {
+closeMatchingMemberIdPopupButton.addEventListener("click", () => {
   document.getElementById("matchingMemberIdPopup").style.display = "none";
 });
 
 // 매칭 채팅방 입장 팝업 제출 버튼 클릭 시
 // 매칭을 통한 채팅방 시작
-document.getElementById("mathingMemberIdForm").addEventListener("submit", (event) => {
+mathingMemberIdForm.addEventListener("submit", (event) => {
   event.preventDefault(); // 기본 제출 동작 방지
   document.getElementById("matchingMemberIdPopup").style.display = "none";
 
   const matchingMemberId = document.getElementById("mathingMemberIdInput").value;
   socket.emit("test-matching-request", { matchingMemberId: matchingMemberId });
-});
-
-// 채팅방 퇴장 시
-function exitChatroom(chatroomUuid) {
-  console.log(`Exit chatroom with UUID: ${chatroomUuid}`);
-
-  // (#14-1) 8080서버에 채팅방 나가기 API 요청
-  exitChatroomApi(chatroomUuid).then(() => {
-    // (#14-2) 채팅방 나가기 API 정상 응답 받음
-    // (#14-3) socket 서버에 exit-chatroom event emit
-    socket.emit("exit-chatroom", { uuid: currentViewingChatroomUuid });
-
-    // (#14-5) 채팅방 영역 초기화
-    resetChatroomDiv(chatroomUuid);
-  });
-}
-
-// 채팅방 내부 메뉴 버튼 생성
-function createChatroomMenuButton(isFriend) {
-  const chatroomHeader = document.querySelector(".column.chatroom h2");
-
-  // 메뉴 버튼 생성
-  const menuButton = document.createElement("button");
-  menuButton.textContent = "⋮";
-  menuButton.id = "menuButton"; // 버튼에 ID 추가
-
-  // 팝업 메뉴 생성
-  const popupMenu = document.createElement("div");
-  popupMenu.id = "popupMenu";
-  popupMenu.className = "popup-menu";
-
-  // 팝업 메뉴 항목 생성
-  //const menuItems = ["채팅방 나가기", "친구 추가", "차단하기", "신고하기", "매너 평가", "비매너 평가"];
-  const ul = document.createElement("ul");
-
-  popupMenu.appendChild(ul);
-
-  // 채팅방 나가기 메뉴 생성
-  createMenuItem(ul, "채팅방 나가기", () => {
-    // 8080서버에 채팅방 나가기 API 요청
-    exitChatroom(currentViewingChatroomUuid);
-  });
-
-  if (isFriend) {
-    // 친구 삭제 메뉴 생성
-    createMenuItem(ul, "친구 삭제", () => {
-      // 8080서버에 친구 삭제 API 요청
-      deleteFriendApi(currentChattingMemberId).then((result) => {
-        // 친구 목록 영역 새로고침
-        const fetchFriendsButton = document.getElementById("fetchFriendsButton");
-        fetchFriendsButton.click(); // 버튼 클릭 이벤트 발생
-
-        // 채팅방 헤더에 상태 element 제거
-        const chatroomHeaderSpan = document.querySelector(".column.chatroom h2 span");
-        if (chatroomHeaderSpan) {
-          chatroomHeaderSpan.textContent = ""; // span 내부 텍스트 초기화
-        }
-      });
-    });
-  } else {
-    // 친구 추가 메뉴 생성
-    createMenuItem(ul, "친구 추가", () => {
-      // 8080서버에 친구 요청 전송 API 요청
-      sendFriendRequestApi(currentChattingMemberId).then((result) => {
-        alert(result.result);
-      });
-    });
-  }
-
-  // 차단하기 메뉴 생성
-  createMenuItem(ul, "차단하기", () => {
-    // 차단 팝업 띄우기
-    // 팝업 요소 선택
-    const blockPopup = document.getElementById("blockPopup");
-    const confirmButton = document.getElementById("confirmBlock");
-    const cancelButton = document.getElementById("cancelBlock");
-
-    blockPopup.style.display = "block";
-
-    // "예" 버튼 클릭 시 실제 이벤트 발생
-    confirmButton.addEventListener("click", () => {
-      // (#15-1) 8080서버에 회원 차단 API 요청
-      blockMemberApi(currentChattingMemberId).then((result) => {
-        // (#15-2) 회원 차단 API 정상 응답 받음
-        blockPopup.style.display = "none"; // 팝업 창 닫기
-        alert(result);
-
-        // (#15-3) socket 서버에 exit-chatroom event emit
-        socket.emit("exit-chatroom", { uuid: currentViewingChatroomUuid });
-
-        // (#15-5) 채팅방 영역 초기화
-        resetChatroomDiv(currentViewingChatroomUuid);
-      });
-      blockPopup.style.display = "none"; // 팝업 창 닫기
-    });
-
-    cancelButton.addEventListener("click", () => {
-      blockPopup.style.display = "none"; // 팝업 창 닫기
-    });
-  });
-
-  // 신고하기 메뉴 생성
-  createMenuItem(ul, "신고하기", () => {
-    // 회원 신고 API 연결
-  });
-
-  // 매너 평가 메뉴 생성
-  createMenuItem(ul, "매너 평가", () => {
-    // 매너 평가 API 연결
-  });
-
-  // 비매너 평가 메뉴 생성
-  createMenuItem(ul, "비매너 평가", () => {
-    // 비매너 평가 API 연결
-  });
-
-  // 채팅방 메뉴 버튼 클릭 시
-  menuButton.addEventListener("click", () => {
-    const isVisible = popupMenu.style.display === "block";
-    popupMenu.style.display = isVisible ? "none" : "block";
-  });
-
-  // 팝업 메뉴 외부를 클릭했을 때 닫히도록 설정
-  document.addEventListener("click", (event) => {
-    if (event.target !== menuButton && !popupMenu.contains(event.target)) {
-      popupMenu.style.display = "none";
-    }
-  });
-
-  // 헤더에 채팅방 메뉴 버튼 및 팝업 메뉴 추가
-  chatroomHeader.appendChild(menuButton);
-  chatroomHeader.appendChild(popupMenu);
-}
-
-// 메뉴 아이템 생성 메소드
-function createMenuItem(ulElement, text, onClick) {
-  const li = document.createElement("li");
-  const button = document.createElement("button");
-  button.className = "menu-item";
-  button.textContent = text;
-
-  button.addEventListener("mouseover", () => {
-    button.style.backgroundColor = "#f1f1f1";
-  });
-
-  button.addEventListener("mouseout", () => {
-    button.style.backgroundColor = "transparent";
-  });
-
-  // 버튼 클릭 시 동작 연결
-  button.addEventListener("click", onClick);
-
-  li.appendChild(button);
-  ulElement.appendChild(li);
-}
-
-// 채팅방 영역 화면 초기화
-function resetChatroomDiv(chatroomUuid) {
-  // chatroom 영역 초기화
-  // 기존 메시지 element 초기화
-  const messagesElement = document.getElementById("messages");
-  messagesElement.innerHTML = "";
-
-  // 채팅방 내부 헤더 초기화
-  const chatroomHeader = document.querySelector(".column.chatroom h2");
-  chatroomHeader.innerHTML = "채팅방";
-
-  // 채팅방 목록 영역에서 해당 채팅방 삭제
-  const chatroomItem = document.querySelector(`.chatroom-item[data-chatroom-uuid="${chatroomUuid}"]`);
-  if (chatroomItem) {
-    chatroomItem.remove();
-  } else {
-    console.log("chatroom delete not found: ", chatroomUuid);
-  }
-
-  // messagesFromThisChatroom array 초기화
-  messagesFromThisChatroom = null;
-
-  // hasNextChat 업데이트
-  hasNextChat = null;
-
-  // 현재 보고 있는 채팅방 uuid, 채팅 중인 memberId 초기화
-  currentViewingChatroomUuid = null;
-  currentChattingMemberId = null;
-
-  // systemFlag 초기화
-  currentSystemFlag = null;
-}
-
-// 채팅방 입장/시작 API 응답으로 채팅방 영역 화면 렌더링 메소드
-function renderChatroomDiv(result) {
-  // 기존 메시지 element 초기화
-  const messagesElement = document.getElementById("messages");
-  const shouldScrollToBottom = messagesElement.scrollTop === messagesElement.scrollHeight - messagesElement.clientHeight;
-
-  messagesElement.innerHTML = "";
-
-  // (#9-7),(#13-7),(#16-8)  API 응답의 chatMessageDtoList에 대해 각 messgae 요소 렌더링
-  if (result.chatMessageList) {
-    result.chatMessageList.chatMessageDtoList.forEach((message) => {
-      const li = document.createElement("li");
-      li.classList.add("message-item");
-
-      if (message.senderId === 0) {
-        // 해당 메시지가 시스템 메시지인 경우
-
-        li.classList.add("system-message");
-        if (message.boardId) {
-          li.setAttribute("data-board-id", message.boardId); // 특정 글로 이동해야 하는 경우, boardId 저장
-          li.addEventListener("click", function () {
-            // 클릭 시 alert 창 띄우기 (원래는 해당 boardId로 게시글 조회 API로 넘어가야 함)
-            alert(`게시판 글 조회 페이지로 이동, board id: ${message.boardId}`);
-          });
-        }
-
-        li.innerHTML = `
-                    <div class="message-content" style = "cursor: pointer;">
-                        <p class="message-text">${message.message}</p>
-                    </div>
-                  `;
-      } else {
-        // 시스템 메시지가 아닌 경우
-        if (message.senderId === memberId) {
-          li.classList.add("mine");
-        }
-        li.innerHTML = `
-                    <div class="message-content">
-                      <img src="${message.senderProfileImg}" alt="Profile Image" width="30" height="30">
-                      <div>
-                        <p class="sender-name">${message.senderName}</p>
-                        <p class="message-text">${message.message}</p>
-                        <p class="message-time">${new Date(message.createdAt).toLocaleString()}</p>
-                      </div>
-                    </div>
-                  `;
-      }
-      messagesElement.appendChild(li);
-    });
-  }
-
-  // Scroll to bottom if was already at bottom before fetching new messages or on initial load
-  if (shouldScrollToBottom || messagesElement.children.length === 0) {
-    messagesElement.scrollTop = messagesElement.scrollHeight;
-  }
-
-  // (#9-8),(#13-8),(#16-9) 채팅방 내부 헤더 렌더링
-  const chatroomHeader = document.querySelector(".column.chatroom h2");
-
-  // 헤더의 profile image, name 업데이트
-  chatroomHeader.innerHTML = `<img src="${result.memberProfileImg}" alt="Profile Image" width="30" height="30" style="vertical-align: middle;">${result.gameName}`;
-
-  // 상대 회원과 내가 친구 관계인 경우에만 상태 text 렌더링
-  if (result.friend) {
-    // 상대 회원이 현재 온라인인 친구 목록에 있는지 여부를 따져서 상태 text 설정
-    const isOnline = onlineFriendMemberIdList.includes(result.memberId);
-    const statusText = isOnline ? "online" : "offline";
-
-    // 상태 element에 class 부여
-    const statusElement = document.createElement("span");
-    statusElement.textContent = `(${statusText})`;
-    statusElement.classList.add(isOnline ? "online" : "offline");
-
-    // 채팅방 헤더에 상태 element 추가
-    chatroomHeader.appendChild(statusElement);
-  }
-
-  // (#9-9),(#13-9),(#16-10) 채팅방 내부 헤더 메뉴 버튼 생성
-  createChatroomMenuButton(result.friend);
-
-  // (#9-10),(#13-10),(#16-11) 채팅방 목록에 읽지 않은 메시지 개수를 0으로 업데이트
-  const chatroomItem = document.querySelector(`.chatroom-item[data-chatroom-uuid="${result.uuid}"] p[data-new-count]`);
-  if (chatroomItem) {
-    chatroomItem.textContent = "0";
-  } else {
-    console.error(`Could not find chatroom item with UUID ${result.uuid} to update new count.`);
-  }
-
-  // (#9-11),(#13-11),(#16-12) 채팅 전송 폼 부분 렌더링
-  // chat-form 부분 추출
-  const chatForm = document.querySelector(".chat-form");
-  const inputField = chatForm.querySelector("input");
-  const sendButton = chatForm.querySelector("button");
-
-  // 상대가 나를 차단한 경우, 채팅 전송 불가하도록 막기
-  if (result.blocked) {
-    // input 필드에 텍스트 설정
-    inputField.value = "대화를 보낼 수 없는 상대입니다.";
-    inputField.disabled = true; // 입력 필드 비활성화
-
-    // 버튼 클릭(폼 전송) 불가 처리
-    sendButton.disabled = true; // 버튼 비활성화
-  } else {
-    // input 필드에 텍스트 초기화
-    inputField.value = "";
-    inputField.disabled = false; // 입력 필드 활성화
-
-    // 버튼 클릭(폼 전송) 가능 처리
-    sendButton.disabled = false; // 버튼 활성화
-  }
-}
-
-// 매칭 성공 후 채팅방 페이지 연결된 경우
-document.addEventListener("DOMContentLoaded", () => {
-  // 세션 스토리지에서 chatroomUuid
-  const chatroomUuid = sessionStorage.getItem("fromMatchPage");
-
-  if (chatroomUuid) {
-    enterChatroom(chatroomUuid);
-
-    // 플래그 삭제 (한 번만 동작하게 함)
-    sessionStorage.removeItem("fromMatchPage");
-  }
-});
-
-// 첫 접속 시 sessionStorage에 플래그 설정
-window.addEventListener("load", () => {
-  // 새로고침 또는 처음 접속 시 sessionStorage에 플래그 설정
-  sessionStorage.setItem("sessionActive", "true");
-});
-
-// 창을 닫을 때 jwt token 스토리지에서 제거
-window.addEventListener("beforeunload", (event) => {
-  // sessionStorage에 값이 없으면 창이 닫히는 것으로 판단
-  if (!sessionStorage.getItem("sessionActive")) {
-    // 창을 닫을 때만 localStorage에서 jwtToken 삭제
-    localStorage.removeItem("jwtToken");
-  }
 });
